[PKG] moves packages to internal
