+++ internal/server/server.go
@@ -5,8 +5,8 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"github.com/xebia/go-exercise/internal/registration/datastore"
-	"github.com/xebia/go-exercise/internal/registration/email"
+	"github.com/xebia/go-exercise/internal/datastore"
+	"github.com/xebia/go-exercise/internal/email"
 	"github.com/xebia/go-exercise/pkg/slogx"
 	"log/slog"
 	"net"
